+++ src/main/java/com/mindmate/mindmate_server/review/service/ReviewService.java
@@ -23,11 +23,5 @@ public interface ReviewService {
 
     ProfileReviewSummaryResponse getProfileReviewSummary(Long profileId);
 
-    Map<String, Integer> getTagCountsByProfileId(Long profileId);
-
-    List<ReviewResponse> getRecentReviewsByUserId(Long userId, int limit);
-
-    Double getAverageRatingByUserId(Long userId);
-
     Review findReviewById(Long reviewId);
 }
